use clap instead of structopt
