+++ Cargo.toml
@@ -31,3 +31,4 @@ colored = "2.0.0"
 fake = { version = "2.5.0", features = ["derive"] }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0.91"
+clap = { version = "4.0.32", features = ["derive"] }
