+++ src/conformance/mod.rs
@@ -1,5 +1,4 @@
 mod basic;
-mod common;
 
 use basic::*;
